chore: centralize socket events
