+++ package.json
@@ -25,6 +25,7 @@
 	},
 	"type": "module",
 	"dependencies": {
+		"dotenv": "^16.0.1",
 		"express": "^4.18.1",
 		"socket.io": "^4.5.1",
 		"socket.io-client": "^4.5.1"
