+++ server/index.js
@@ -5,6 +5,7 @@ import { createServer } from 'http';
 import { Server } from 'socket.io';
 
 import { handler } from '../build/handler.js';
+import { socketEvents } from './socket.js';
 
 const port = process.env.PORT || 3000;
 const app = express();
@@ -12,9 +13,8 @@ const server = createServer(app);
 
 const io = new Server(server);
 
-// TODO extract these
 io.on('connection', (socket) => {
-	socket.emit('serverEvent', 'hello world 👋');
+	socketEvents(io, socket);
 });
 
 app.use(handler);
