+++ src/routes/__layout.svelte
@@ -7,19 +7,16 @@
 
 	// set all users
 	$socket.on('setUsers', (data) => {
-		console.log('set users', data);
 		$users = data;
 	});
 
 	// set user id (self)
 	$socket.on('setUser', (data) => {
-		console.log('set user', data);
 		$user = data;
 	});
 
 	// user joins room
 	$socket.on('joinRoom', (data) => {
-		console.log('join room', data);
 		$users = data.users;
 	});
